# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "AI Tutor Backend API",
    "version": "1.0.0",
    "description": "REST API for authentication, dialogs, files, quiz generation, and user management."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "QuizQuestionResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "question": {
            "type": "string"
          },
          "variants": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "id",
          "question",
          "variants"
        ]
      },
      "QuizResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "test_name": {
            "type": "string"
          },
          "questions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuizQuestionResponse"
            }
          }
        },
        "required": [
          "id",
          "test_name",
          "questions"
        ]
      },
      "QuizAnswerRequest": {
        "type": "object",
        "properties": {
          "questionId": {
            "type": "integer",
            "format": "int64"
          },
          "answer": {
            "type": "string"
          }
        },
        "required": [
          "questionId"
        ]
      },
      "QuizScoreRequest": {
        "type": "object",
        "properties": {
          "answers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuizAnswerRequest"
            }
          }
        },
        "required": [
          "answers"
        ]
      },
      "QuizQuestionScoreResponse": {
        "type": "object",
        "properties": {
          "questionId": {
            "type": "integer",
            "format": "int64"
          },
          "selectedAnswer": {
            "type": "string"
          },
          "correctAnswer": {
            "type": "string"
          },
          "correct": {
            "type": "boolean"
          }
        },
        "required": [
          "questionId",
          "correctAnswer",
          "correct"
        ]
      },
      "QuizScoreResponse": {
        "type": "object",
        "properties": {
          "quizId": {
            "type": "integer",
            "format": "int64"
          },
          "totalQuestions": {
            "type": "integer"
          },
          "correctAnswers": {
            "type": "integer"
          },
          "score": {
            "type": "number",
            "format": "double"
          },
          "questions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuizQuestionScoreResponse"
            }
          }
        },
        "required": [
          "quizId",
          "totalQuestions",
          "correctAnswers",
          "score",
          "questions"
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "integer"
          },
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "path": {
            "type": "string"
          }
        },
        "required": [
          "timestamp",
          "status",
          "error",
          "message",
          "path"
        ]
      }
    }
  },
  "paths": {
...
    "/api/quiz": {
      "post": {
        "summary": "Generate and persist quiz for a dialog",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "dialogId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "questionsCount",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 20
            },
            "description": "Количество вопросов в квизе."
          }
        ],
        "responses": {
          "201": {
            "description": "Quiz generated and saved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuizResponse"
                }
              }
            }
          },
          "400": {
            "description": "Некорректное количество вопросов в квизе",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Dialog not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Get all quizzes for a dialog",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "dialogId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Quizzes retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/QuizResponse"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Dialog not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/quiz/{quizId}": {
      "get": {
        "summary": "Get a single quiz by id within a dialog",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "quizId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "dialogId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Quiz retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuizResponse"
                }
              }
            }
          },
          "404": {
            "description": "Quiz or dialog not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/quiz/{quizId}/score": {
      "post": {
        "summary": "Score submitted quiz answers",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "quizId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuizScoreRequest"
              },
              "example": {
                "answers": [
                  {
                    "questionId": 1,
                    "answer": "Language"
                  },
                  {
                    "questionId": 2,
                    "answer": "Operating system"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Quiz scored",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuizScoreResponse"
                }
              }
            }
          },
          "404": {
            "description": "Quiz not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  }
}